requirements.txt
amqp==5.0.6
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
currency-open-exchange==0.0.6
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.44.0
django-celery==3.3.1
django-celery-beat==2.2.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-exchange==0.7.4
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.3.0
django-timezone-field==4.1.2
djangorestframework==3.11.0
filelock==3.0.12
forex-python==1.5
google-currency==1.0.9
gunicorn==20.1.0
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
openexchangerates==0.1.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==7.1.2
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyCurrency-Converter==1.0
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
stripe==2.56.0
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
whitenoise==5.2.0
